pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "harnessa"
version = "0.1.0"
description = "GAN-inspired multi-agent harness framework for LLM evaluation"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0",
    "litellm>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
harnessa = "harnessa.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
